requirements.txt
aiohttp
asyncio
beautifulsoup4
numpy
pandas
Pillow
requests
scikit-learn
webcolors
